chore: rename private-mode-switch to private-mode-checkbox
Refs: #289
