+++ src/components/pages/account-page/current-user-info/private-mode-switch/private-mode-checkbox/index.tsx
